Move testParquetOnlyNullsRowGroupPruning to io.trino.testing
